xarbb/xaradmin/delete.php: 	#  added changes by Georg Schnattinger xarbb/xaradmin/main.php: 	#  added changes by Georg Schnattinger xarbb/xaradmin/modify.php: 	#  added changes by Georg Schnattinger xarbb/xaradmin/new.php: 	#  added changes by Georg Schnattinger xarbb/xaradmin/view.php: 	#  added changes by Georg Schnattinger xarbb/xaradminapi/create.php: 	#  added changes by Georg Schnattinger xarbb/xaradminapi/delete.php: 	#  added changes by Georg Schnattinger xarbb/xaradminapi/deletealltopics.php: 	#  added changes by Georg Schnattinger xarbb/xaradminapi/deletetopics.php: 	#  added changes by Georg Schnattinger xarbb/xaradminapi/getmenulinks.php: 	#  added changes by Georg Schnattinger xarbb/xaradminapi/update.php: 	#  added changes by Georg Schnattinger xarbb/xarinit.php: 	#  added changes by Georg Schnattinger xarbb/xartemplates/user-newtopic.xd: 	#  added changes by Georg Schnattinger xarbb/xartemplates/user-viewforum.xd: 	#  added changes by Georg Schnattinger xarbb/xartemplates/user-viewtopic.xd: 	#  added changes by Georg Schnattinger xarbb/xaruser/main.php: 	#  added changes by Georg Schnattinger xarbb/xaruser/newreply.php: 	#  added changes by Georg Schnattinger xarbb/xaruser/newtopic.php: 	#  added changes by Georg Schnattinger xarbb/xaruser/viewforum.php: 	#  added changes by Georg Schnattinger xarbb/xaruser/viewforumrss.php: 	#  added changes by Georg Schnattinger xarbb/xaruser/viewtopic.php: 	#  added changes by Georg Schnattinger xarbb/xaruserapi/countposts.php: 	#  added changes by Georg Schnattinger xarbb/xaruserapi/createtopic.php: 	#  added changes by Georg Schnattinger xarbb/xaruserapi/getallforums.php: 	#  added changes by Georg Schnattinger xarbb/xaruserapi/getalltopics.php: 	#  added changes by Georg Schnattinger xarbb/xaruserapi/getforum.php: 	#  added changes by Georg Schnattinger xarbb/xaruserapi/gettopic.php: 	#  added changes by Georg Schnattinger xarbb/xaruserapi/updateforumview.php: 	#  added changes by Georg Schnattinger xarbb/xaruserapi/updatetopicsview.php: 	#  added changes by Georg Schnattinger xarbb/xaradminapi/deleteallreplies.php: 	#  BitKeeper file C:/cygwin/home/phpdev/www/dev/modules/xarbb/xaradminapi/deleteallreplies.php xarbb/xaradminapi/deletereplies.php: 	#  BitKeeper file C:/cygwin/home/phpdev/www/dev/modules/xarbb/xaradminapi/deletereplies.php xarbb/xartemplates/user-deletereply.xd: 	#  BitKeeper file C:/cygwin/home/phpdev/www/dev/modules/xarbb/xartemplates/user-deletereply.xd xarbb/xartemplates/user-deletetopic.xd: 	#  BitKeeper file C:/cygwin/home/phpdev/www/dev/modules/xarbb/xartemplates/user-deletetopic.xd xarbb/xaruser/deletereply.php: 	#  BitKeeper file C:/cygwin/home/phpdev/www/dev/modules/xarbb/xaruser/deletereply.php xarbb/xaruser/deletetopic.php: 	#  BitKeeper file C:/cygwin/home/phpdev/www/dev/modules/xarbb/xaruser/deletetopic.php xarbb/xaruserapi/getitemtypes.php: 	#  BitKeeper file C:/cygwin/home/phpdev/www/dev/modules/xarbb/xaruserapi/getitemtypes.php xarbb/xaruserapi/updatetopic.php: 	#  BitKeeper file C:/cygwin/home/phpdev/www/dev/modules/xarbb/xaruserapi/updatetopic.php

+++ xarbb/xaradminapi/deletetopics.php
@@ -2,7 +2,7 @@
 
 /**
  * delete a forum
- * @param $args['fid'] ID of the forum
+ * @param $args['tids'] Array( IDs ) of the forum   or $args['tid'] ID
  * @returns bool
  * @return true on success, false on failure
  */
@@ -14,16 +14,17 @@ function xarbb_adminapi_deletetopics($args)
     extract($args);
 
     // Argument check
-    if (!isset($fid)) {
-        $msg = xarML('Invalid Parameter Count',
+    if ( (!isset($tids) || !is_array($tids) || count($tids) == 0) &&
+    	 (!isset($tid) || !($tid > 0)) ) {
+        $msg = xarML('Invalid Parameter count',
                     join(', ',$invalid), 'admin', 'delete', 'xarbb');
         xarExceptionSet(XAR_SYSTEM_EXCEPTION, 'BAD_PARAM',
                        new SystemException($msg));
         return;
     }
 
-    // Security Check
-    if(!xarSecurityCheck('DeletexarBB')) return;
+    if(!isset($tids))
+    	$tids = Array($tid);
 
     // Get datbase setup
     list($dbconn) = xarDBGetConn();
@@ -31,15 +32,29 @@ function xarbb_adminapi_deletetopics($args)
 
     $xbbtopicstable = $xartable['xbbtopics'];
 
-    // Delete the item
-    $query = "DELETE FROM $xbbtopicstable
-              WHERE xar_fid = " . xarVarPrepForStore($fid);
-    $result =& $dbconn->Execute($query);
-    if (!$result) return;
+    foreach($tids as $tid)	{
+    	// get forum id
+        if(!$topic = xarModAPIFunc('xarbb','user','gettopic',array('tid' => $tid))) return;
+		$fid = $topic['fid'];
 
-    // Let any hooks know that we have deleted a link
-// should call item delete hooks for each topic, really
-//    xarModCallHooks('item', 'delete', $fid, '');
+	    // Item Specific Security Check
+	    if(!xarSecurityCheck('ModxarBB',1,'Forum',"$fid:All")) continue;
+
+		// Delete comments
+        if(!xarModAPIFunc("xarbb","admin","deleteallreplies",array(
+        			"tid" => $tid
+				))) return;
+	    // Delete the item
+	    $query = "DELETE FROM $xbbtopicstable
+	              WHERE xar_tid = $tid";
+	    $result =& $dbconn->Execute($query);
+	    if (!$result) return;
+
+        // Let any hooks know that we have deleted a topic
+        $args['module'] = 'xarbb';
+	    $args['itemtype'] = 2; // topic
+	    xarModCallHooks('item', 'delete', $tid, $args);
+	}
 
     // Let the calling process know that we have finished successfully
     return true;
